refactor(azure): modularize Azure into dedicated package and clean cloud dispatcher
- Added discovr/azure/__init__.py to re-export AzureDiscovery, AzureExporter, AzureReporter
- Refactored discovr/cloud.py to act only as dispatcher for non-Azure providers (GCP, AWS placeholder)
- Removed Azure import clutter from cloud dispatcher
- Ensures Azure discovery is handled cleanly via discovr.azure package
